requirements.txt
flask>=3.0.0
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
pandas>=2.2.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
numpy>=1.26.0
